refactor(compiler-sfc): split more logic
