+++ packages/compiler-sfc/src/script/utils.ts
@@ -9,7 +9,7 @@ export function resolveObjectKey(node: Node, computed: boolean) {
   switch (node.type) {
     case 'StringLiteral':
     case 'NumericLiteral':
-      return node.value
+      return String(node.value)
     case 'Identifier':
       if (!computed) return node.name
   }
